Final project
The research question is how e-voting can be utilized to reduce government money, while the motivation is to create a secure and credible electoral system. The aim of the study is to improve security and speed up the counting of votes using advances in blockchain technology. The research objectives include problem identification, analysis of how technology is used to conduct elections, research design to create an efficient system that can reduce election expenses, and evaluation to ensure that the achieved methodology is implemented in a manner that minimizes costs and maintains confidentiality.